feat: ContextProvider for state management
