+++ src/components/RenderForm.js
@@ -1,5 +1,4 @@
-import React, { createRef } from 'react'
-import PropTypes from 'prop-types'
+import React, { createRef, useContext } from 'react'
 import Form from '@rjsf/core'
 import { customizeValidator } from '@rjsf/validator-ajv8'
 import { widgets } from '../custom-ui/CustomWidgets'
@@ -8,20 +7,16 @@ import Config from '../utils/config'
 import { saveToJSONFile } from '../utils/helpers/file.helpers'
 import { toast } from 'react-toastify'
 import { Button } from './layout/inputs'
+import { SchemaContext } from '../contexts/schema.context'
 
-function RenderForm (props) {
-  /*
-  Method to read and render a form based on the user-selected schema
-  Arguments:
-    props.schemaType (string) = the selected schema type
-    props.schema (object) = the selected schema
-    props.formData (object) = the form data
-  Returns:
-    Form object
-  */
-  const { schemaType, schema, formData } = props
+/**
+ * Component to read and render a form based on the user-selected schema.
+ * Uses SchemaContext to get states (e.g. schema, formData) to render the form. for the form)
+ */
+function RenderForm () {
+  const { selectedSchemaType, schema, formData } = useContext(SchemaContext)
   const formRef = createRef()
-  const validator = customizeValidator(Config.ajvOptions)
+  const validator = customizeValidator(Config.AJV_OPTIONS)
 
   /**
    * Event handler called after rjsf form validation
@@ -37,7 +32,7 @@ function RenderForm (props) {
    * @param {Event} event The form event
    */
   async function saveFileOnSubmit (event) {
-    saveToJSONFile(event.formData, schemaType)
+    saveToJSONFile(event.formData, selectedSchemaType)
   }
 
   /**
@@ -77,12 +72,6 @@ function RenderForm (props) {
   } else {
     return (<div>Please select a schema from the dropdown above or autofill data from an existing file.</div>)
   }
-}
-
-RenderForm.propTypes = {
-  schemaType: PropTypes.string.isRequired,
-  schema: PropTypes.object,
-  formData: PropTypes.object
 }
 
 export default RenderForm
