+++ src/tests/components/SchemaToolbar.test.js
@@ -5,21 +5,34 @@ import SCHEMA_LIST from '../resources/schema-links/full-parsed-filtered.json'
 import SORTED_VERSION_LIST_INSTRUMENT from '../resources/sorted-version-list-instrument.json'
 import { toast } from 'react-toastify'
 import Help from '../../components/Help'
+import { SchemaContext } from '../../contexts/schema.context'
 
 const NULL_CALLBACK = () => { }
 jest.mock('react-toastify', () => ({ toast: jest.fn() }))
 jest.mock('../../components/Help', () => jest.fn())
 
+/**
+ * Helper function to render SchemaToolbar component with test context
+ * @param {*} context - Context value to used in the test
+ */
+const renderWithContext = (context) => {
+  render(
+    <SchemaContext.Provider value={context}>
+      <SchemaToolbar/>
+    </SchemaContext.Provider>
+  )
+}
+
 describe('SchemaToolbar component', () => {
   it('renders appropriate inputs on default', () => {
-    render(<SchemaToolbar
-      updateSelectedSchemaType={NULL_CALLBACK}
-      updateSelectedSchemaVersion={NULL_CALLBACK}
-      selectedSchemaType=''
-      selectedSchemaPath=''
-      schemaList={SCHEMA_LIST}
-      handleRehydrate={NULL_CALLBACK}
-    />)
+    renderWithContext({
+      updateSelectedSchemaType: NULL_CALLBACK,
+      updateSelectedSchemaVersion: NULL_CALLBACK,
+      selectedSchemaType: '',
+      selectedSchemaPath: '',
+      schemaList: SCHEMA_LIST,
+      handleRehydrate: NULL_CALLBACK
+    })
     expect(screen.getByTitle('Select schema')).toBeInTheDocument()
     expect(screen.getByTitle('Select version')).toBeInTheDocument()
     expect(screen.getByTitle('Autofill with existing data from local file')).toBeInTheDocument()
@@ -34,14 +47,14 @@ describe('SchemaToolbar component', () => {
     const mockTypeCallback = jest.fn()
     const schemaType = 'subject'
     const newSchemaType = 'instrument'
-    render(<SchemaToolbar
-      updateSelectedSchemaType={mockTypeCallback}
-      updateSelectedSchemaVersion={NULL_CALLBACK}
-      selectedSchemaType={schemaType}
-      selectedSchemaPath=''
-      schemaList={SCHEMA_LIST}
-      handleRehydrate={NULL_CALLBACK}
-    />)
+    renderWithContext({
+      updateSelectedSchemaType: mockTypeCallback,
+      updateSelectedSchemaVersion: NULL_CALLBACK,
+      selectedSchemaType: schemaType,
+      selectedSchemaPath: '',
+      schemaList: SCHEMA_LIST,
+      handleRehydrate: NULL_CALLBACK
+    })
     fireEvent.change(screen.getByTitle('Select schema'), { target: { value: newSchemaType } })
     expect(screen.getByTitle('Select version')).toBeEnabled()
     expect(mockTypeCallback).toHaveBeenCalledWith(newSchemaType)
@@ -49,14 +62,14 @@ describe('SchemaToolbar component', () => {
 
   it('has schema versions sorted by latest-first semantic version', () => {
     const schemaType = 'instrument'
-    render(<SchemaToolbar
-      updateSelectedSchemaType={NULL_CALLBACK}
-      updateSelectedSchemaVersion={NULL_CALLBACK}
-      selectedSchemaType={schemaType}
-      selectedSchemaPath=''
-      schemaList={SCHEMA_LIST}
-      handleRehydrate={NULL_CALLBACK}
-    />)
+    renderWithContext({
+      updateSelectedSchemaType: NULL_CALLBACK,
+      updateSelectedSchemaVersion: NULL_CALLBACK,
+      selectedSchemaType: schemaType,
+      selectedSchemaPath: '',
+      schemaList: SCHEMA_LIST,
+      handleRehydrate: NULL_CALLBACK
+    })
     const schemaVersionsList = [...screen.getByTitle('Select version').options].map((option) => option.text)
     expect(schemaVersionsList).toStrictEqual(SORTED_VERSION_LIST_INSTRUMENT)
   })
@@ -65,28 +78,28 @@ describe('SchemaToolbar component', () => {
     const mockVersionCallback = jest.fn()
     const schemaType = 'instrument'
     const schemaPath = 'schemas/instrument/0.10.0/instrument_schema.json'
-    render(<SchemaToolbar
-      updateSelectedSchemaType={NULL_CALLBACK}
-      updateSelectedSchemaVersion={mockVersionCallback}
-      selectedSchemaType={schemaType}
-      selectedSchemaPath=''
-      schemaList={SCHEMA_LIST}
-      handleRehydrate={NULL_CALLBACK}
-    />)
+    renderWithContext({
+      updateSelectedSchemaType: NULL_CALLBACK,
+      updateSelectedSchemaVersion: mockVersionCallback,
+      selectedSchemaType: schemaType,
+      selectedSchemaPath: '',
+      schemaList: SCHEMA_LIST,
+      handleRehydrate: NULL_CALLBACK
+    })
     fireEvent.change(screen.getByTitle('Select version'), { target: { value: schemaPath } })
     expect(mockVersionCallback).toHaveBeenCalledWith(schemaPath)
   })
 
   it('calls handleRehydrate when autofill button is clicked', () => {
     const mockRehydrateCallback = jest.fn()
-    render(<SchemaToolbar
-      updateSelectedSchemaType={NULL_CALLBACK}
-      updateSelectedSchemaVersion={NULL_CALLBACK}
-      selectedSchemaType=''
-      selectedSchemaPath=''
-      schemaList={SCHEMA_LIST}
-      handleRehydrate={mockRehydrateCallback}
-    />)
+    renderWithContext({
+      updateSelectedSchemaType: NULL_CALLBACK,
+      updateSelectedSchemaVersion: NULL_CALLBACK,
+      selectedSchemaType: '',
+      selectedSchemaPath: '',
+      schemaList: SCHEMA_LIST,
+      handleRehydrate: mockRehydrateCallback
+    })
     fireEvent.click(screen.getByTitle('Autofill with existing data from local file'))
     expect(mockRehydrateCallback).toHaveBeenCalled()
   })
@@ -96,14 +109,14 @@ describe('SchemaToolbar component', () => {
       toastId: 'help-toast',
       autoClose: false
     }
-    render(<SchemaToolbar
-      updateSelectedSchemaType={NULL_CALLBACK}
-      updateSelectedSchemaVersion={NULL_CALLBACK}
-      selectedSchemaType=''
-      selectedSchemaPath=''
-      schemaList={SCHEMA_LIST}
-      handleRehydrate={NULL_CALLBACK}
-    />)
+    renderWithContext({
+      updateSelectedSchemaType: NULL_CALLBACK,
+      updateSelectedSchemaVersion: NULL_CALLBACK,
+      selectedSchemaType: '',
+      selectedSchemaPath: '',
+      schemaList: SCHEMA_LIST,
+      handleRehydrate: NULL_CALLBACK
+    })
     fireEvent.click(screen.getByTitle('Get help'))
     expect(toast).toHaveBeenCalledWith((<Help />), expectedToastParams)
   })
